chore: annoying onboarding state issue
