+++ client/src-tauri/src/structs/mod.rs
@@ -1,2 +1,1 @@
 pub(crate) mod app_state;
-pub(crate) mod onboarding_state;
